// web/src/api/generated/models/catalogSellableModifierGroupResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * POS Cafe API
 * Backend API for Cafe Point of Sale System built with Vertical Slice Architecture and Idiomatic Go.
 * OpenAPI spec version: 1.0
 */
import type { CatalogSellableModifierOptionResponse } from './catalogSellableModifierOptionResponse';

export interface CatalogSellableModifierGroupResponse {
  default_option_ids?: string[];
  id?: string;
  max_selections?: number;
  min_selections?: number;
  name?: string;
  options?: CatalogSellableModifierOptionResponse[];
}